Define legacy 'jpi' name in file instead of build.gradle
Newer versions of com.gradle.plugin-publish and maven-publish
create implicit configuration around each of the plugins defined
in the `gradlePlugins` block of the build.gradle file. By moving
it to just a classpath resource, we can continue to support the
legacy plugin application just as well without the extra behavior.