Migrate JpiPomCustomizerSpec to GradleTestKit
This unblocks the move to Gradle 5.2

One benefit of switching to the maven-publish plugin is pom
generation now happens in a dedicated task. We can easily test the
pom that gradle generates with GradleTestKit, without relying on
anything internal.

This test previously relied on internal gradle classes and project
evaluation, which has broken on two gradle upgrades in the past
year.