Put work directory inside project directory
Previously the rootDir was used as parent for the work directory. This
led to the situation where a subproject containing a jenkins plugin put
the work directory at the top of the project hierarchy. This becomes
especially problematic when several jenkins plugins are managed from one
root projects.

This change uses `project.projectDir` instead which will be the same as
`project.rootDir` for projects without sub projcts but will be the
actual project directory for nested projects.